requirements.txt
python-dotenv
yandex-tracker-client
scikit-learn
pandas
pyTelegramBotAPI